+++ apps/miranum-modeler/src/adapter/in/commands.ts
@@ -1,11 +1,15 @@
-import { commands } from "vscode";
+import { commands, env, Uri, workspace } from "vscode";
 import { inject, singleton } from "tsyringe";
 
 import { getContext } from "@miranum-ide/vscode/miranum-vscode";
 import {
+    GetDiagramAsSvgInPort,
+    GetDocumentInPort,
     OpenLoggingConsoleInPort,
     ToggleTextEditorInPort,
 } from "../../application/ports/in";
+import { subscribeToMessageEvent } from "../out";
+import { GetDiagramAsSVGCommand } from "@miranum-ide/vscode/miranum-vscode-webview";
 
 @singleton()
 export class VsCodeToggleTextEditorCommand {
@@ -50,5 +54,54 @@ export class VsCodeOpenLoggingConsoleCommand {
 
     show(): void {
         this.openLoggingConsoleInPort.open();
+    }
+}
+
+@singleton()
+export class VsCodeDiagramAsSvgCommand {
+    constructor(
+        @inject("CopyDiagramAsSvgCommand") writeToClipboard: string,
+        @inject("SaveDiagramAsSvgCommand") writeToFile: string,
+        @inject("GetDiagramAsSvgInPort")
+        private readonly getDiagramAsSvgInPort: GetDiagramAsSvgInPort,
+        @inject("GetDocumentInPort")
+        private readonly getDocumentInPort: GetDocumentInPort,
+    ) {
+        const context = getContext();
+
+        context.subscriptions.push(
+            commands.registerCommand(writeToClipboard, this.writeToClipboard, this),
+            commands.registerCommand(writeToFile, this.writeToFile, this),
+        );
+    }
+
+    writeToClipboard(): void {
+        this.getDiagramAsSvgInPort.getSvg();
+
+        subscribeToMessageEvent((message) => {
+            if (message.type === "GetDiagramAsSVGCommand") {
+                const command = message as GetDiagramAsSVGCommand;
+                if (command.svg && command.svg.length > 0) {
+                    env.clipboard.writeText(command.svg);
+                }
+            }
+        });
+    }
+
+    writeToFile(): void {
+        this.getDiagramAsSvgInPort.getSvg();
+
+        subscribeToMessageEvent((message) => {
+            if (message.type === "GetDiagramAsSVGCommand") {
+                const command = message as GetDiagramAsSVGCommand;
+                if (command.svg && command.svg.length > 0) {
+                    const filePath = this.getDocumentInPort
+                        .getPath()
+                        .replace("bpmn", "svg");
+
+                    workspace.fs.writeFile(Uri.file(filePath), Buffer.from(command.svg));
+                }
+            }
+        });
     }
 }